requirements.txt
Django==3.2.6
pymongo==3.12.0
djangorestframework==3.12.4
